Fix minimize crash

+++ pilcrow/engine/core/settings/WindowSettings.h
@@ -3,6 +3,6 @@
 #pragma once
 
 // init in main
-extern float g_InitialWindowWidth;
-extern float g_InitialWindowHeight;
+extern unsigned int g_InitialWindowWidth;
+extern unsigned int g_InitialWindowHeight;
 extern bool  g_StartFullscreen;
